refactor : 뒤로 가기 방지 로직 hook으로 추상화
